dom/base/test/jsmodules/importmaps/multiple/test_importMap_between_dynamic_imports.html

<!DOCTYPE html>
<head>
  <meta charset=utf-8>
  <title>Test import map's entry is ignored if the specifier has been resolved.</title>
</head>
<body onload='testLoaded()'>
<script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

<script>
  import("./module_simpleExport.mjs").then((ns) => {
    ok(ns.x == 42, 'Check imported value result before registering an import map: ' + ns.x);
  }).catch((e) => {
    ok(false, "throws " + e);
  })
</script>

<!-- module_simpleExport.mjs has been resolved, so the entry -->
<!-- "./module_simpleExport.mjs" in the import map won't be added. -->

<script type="importmap" onerror='importMapError()'>
{
  "imports": {
    "./module_simpleExport.mjs": "./scope1/module_simpleExport.mjs"
  }
}
</script>

<script>
  SimpleTest.waitForExplicitFinish();

  let hasError = false;
  // eslint-disable-next-line no-unused-vars
  function importMapError() {
    hasError = true;
  }

  // eslint-disable-next-line no-unused-vars
  function testLoaded() {
    // doing import() again should get the same result, regardless the import map.
    import("./module_simpleExport.mjs").then((ns) => {
      ok(ns.x == 42, 'Check simple imported value result after registering an import map: ' + ns.x);
      ok(!hasError, "onerror of the import map shouldn't be called.");
    }).catch((e) => {
      ok(false, "throws " + e);
    }).then(() => {
      SimpleTest.finish();
    });
  }
</script>
</body>
